♻️ Refactor: 맵핑 공통함수 DrugMapper로 분리
